add method to handle recipe creation
